越野  程序

+++ MCU/Travel_Car/HARDWARE/SENSOR8/sensor8.h
@@ -4,6 +4,6 @@
 
 void Sensor_Init(void);
 int Read_Data1(void);
-void Read_Data(uint16_t *Data);		
+void Read_Data(int *Data);		
 
 #endif
